SPEAD header now matches specification and a subtle but nasty bug fix
The SPEAD header had an incorrect magic number 0x83 instead of 0x53 and did not have
the correct fileds for item pointer width (3) and heap addr width (5).

It was found that changing the magic number broke SPEAD and this was eventually
tracked down to a subtle back related to the new magic number not having the MSB set.

In some cases the header was accidently decoded as an item and interpreted as being
a direct addressed item which broke the packet linkage further down. Basically
just an off by one error but darn subtle....